Still borken badly
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@4098 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/int/linear.hh
@@ -1218,7 +1218,7 @@ namespace Gecode { namespace Int { namespace Linear {
    * Rewrites \a r into IRT_EQ, IRT_NQ, IRT_LQ, IRT_GQ and changes
    * \a c accordingly.
    */
-  void normalize(IntRelType& r, int& c);
+  void normalize(IntRelType& r, double& c);
 
   /**
    * \brief Post propagator for linear constraint over integers
